python/user-activity-analysis/requirements.txt
oracledb==3.1.1
sqlalchemy==2.0.41
requests==2.32.3
python-dotenv==1.1.0
pyyaml==6.0.2
